+++ src/database/InMemoryStorage.ts
@@ -6,9 +6,9 @@ class InMemoryStorage<T> {
   }
 
   public async findBy(key: keyof T, value: any): Promise<T | undefined> {
-    const item = this.items.find((item) => item[key] === value);
+    const findedItem = this.items.find((item) => item[key] === value);
 
-    return item;
+    return findedItem;
   }
 
   public async add(item: T): Promise<void> {
@@ -17,11 +17,11 @@ class InMemoryStorage<T> {
 
   public async update(key: keyof T, value: any, updatedItem: T): Promise<T> {
     const newItems = this.items.map((item) => {
-        if (item[key] === value) {
-            return updatedItem;
-        }
+      if (item[key] === value) {
+        return updatedItem;
+      }
 
-        return item;
+      return item;
     });
 
     this.items = newItems;
@@ -34,6 +34,6 @@ class InMemoryStorage<T> {
 
     this.items = newItems;
   }
-};
+}
 
 export default InMemoryStorage;
