--- src/framework/RequestType.ts
@@ -1,8 +1,0 @@
-import { IncomingMessage } from 'http';
-
-interface RequestType extends IncomingMessage {
-  body: any;
-  id: string;
-}
-
-export default RequestType;
