--- src/framework/ServerResponseType.ts
@@ -1,7 +1,0 @@
-import { ServerResponse } from 'http';
-
-interface ServerResponseType extends ServerResponse {
-  send: (status: number, data: any) => void;
-}
-
-export default ServerResponseType;
