--- src/framework/requestMethods.ts
@@ -1,8 +1,0 @@
-enum RequestMethods {
-  GET = 'GET',
-  POST = 'POST',
-  PUT = 'PUT',
-  DELETE = 'DELETE',
-}
-
-export default RequestMethods;
