+++ src/index.ts
@@ -5,6 +5,7 @@ import userRouter from './components/user/userRouter';
 import jsonParser from './framework/middlewares/jsonParser';
 import bodyParser from './framework/middlewares/bodyParser';
 import urlParser from './framework/middlewares/urlParser';
+import { handleExceptionError } from './framework/middlewares/errors/handleErrors';
 
 dotenv.config();
 
@@ -15,6 +16,7 @@ const app = new Application();
 app.use(jsonParser);
 app.use(bodyParser);
 app.use(urlParser);
+app.use(handleExceptionError);
 
 app.addRouter(userRouter);
