+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "module": "ES6",
     "moduleResolution": "Node",
-    "target": "ES5",
+    "target": "ES6",
     "allowSyntheticDefaultImports": true,
     "outDir": "dist",
     "strictNullChecks": true,
